pyproject.toml
[project]
name = "flight-search-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.19.0",
    "psycopg2-binary>=2.9.11",
    "sentence-transformers>=5.1.2",
]
